% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment analysis for network module interpretation}
\usage{
enrichment(
  genelist,
  background,
  db,
  species,
  Num = 2,
  test_method = "fisher",
  adjustment = "fdr"
)
}
\arguments{
\item{genelist}{The list of gene that is used for enrichment analysis. Needs to be entrez id in character form}

\item{background}{The background gene set user specify, it should be in the same form as the genelist}

\item{db}{Database for enrichment, either KEGG or GO}

\item{species}{The species that experiment data is dealing with}

\item{Num}{At least what number of genes should be in a specific pathway}

\item{test_method}{Either can be Fisher's exact test or Chi-square test}

\item{adjustment}{P-value adjustment method, default set to FDR}
}
\description{
This is the enrichment analysis that can be done through either fisher exact test or chi-square test
by changing the background gene set into the one that we specify, and obtain the enrichment pathways
with significance either through KEGG or GO
}
