% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Enrichment.R
\name{evaluation}
\alias{evaluation}
\title{Grid search evaluation method that output the final scaling factor}
\usage{
evaluation(Error)
}
\arguments{
\item{Error}{The error data output from get_error() function}
}
\description{
Grid search evaluation method that output the final scaling factor
}
