% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetTools.R
\name{Extract_Information}
\alias{Extract_Information}
\title{Extracting Omics modules, heatmap and other grid search related measures}
\usage{
Extract_Information(CVDirs, grids, phenos, cutting)
}
\arguments{
\item{CVDirs}{result directory based on the previous SmCCNet running directory, which
is given by the output of overall_running() function}

\item{grids}{this is the number of grids that are used for the grid search, which is given by length(factor_1)}

\item{phenos}{the name of the phenotype}

\item{cutting}{a vector of cutting edge we are using, which should be consistent with overall_running()}
}
\value{
summary results in the result directory
}
\description{
Extracting Omics modules, heatmap and other grid search related measures
}
