% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_Enrichment.R
\name{get_error}
\alias{get_error}
\title{Obtain CV error for all grids}
\usage{
get_error(Dir, grids, subSamp)
}
\arguments{
\item{Dir}{The working directory from previous running}

\item{grids}{Number of grids that are using from the previous running}

\item{subSamp}{The subsampling number used in previous running, must be the same since this is the
call that read in the data file}
}
\description{
Obtain CV error for all grids
}
