% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAdiagTools.R
\name{GetFuncGrp}
\alias{GetFuncGrp}
\title{(INTERNAL)}
\usage{
GetFuncGrp(
  sccaFile,
  SaveResultFile,
  SavePlotFile = NULL,
  Cutoff = 1,
  GetPerformance = FALSE,
  Signal = NULL
)
}
\description{
(INTERNAL)
}
\keyword{internal}
