% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetTools.R
\name{overall_running}
\alias{overall_running}
\title{Running overall SmCCNet based on previous cross-validation result}
\usage{
overall_running(
  dir,
  CVDirs,
  X1,
  X2,
  Y,
  factor_1,
  factor_2,
  factor_3,
  pheno,
  cutting = c(0, 0.01, 0.3),
  subSamp = 1000,
  s1 = 0.7,
  s2 = 0.9,
  K = 5,
  phenos
)
}
\arguments{
\item{dir}{working directory}

\item{CVDirs}{result directory}

\item{X1}{first dataset}

\item{X2}{second dataset}

\item{Y}{phenotype}

\item{factor_1}{first scaling factor}

\item{factor_2}{second scaling factor}

\item{factor_3}{third scaling factor}

\item{pheno}{phenotype name}

\item{cutting}{a sequence of cutting edge}

\item{subSamp}{number of subsampling}

\item{s1}{first subsampling parameter}

\item{s2}{second subsampling parameter}

\item{K}{cross-validation parameter}

\item{phenos}{name of phenotype}
}
\value{
SmCCNet result stored in the local directory
}
\description{
Running overall SmCCNet based on previous cross-validation result
}
