% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAdiagTools.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{(INTERNAL)}
\usage{
PlotNetwork(
  EdgeMatrix,
  NodeIdx,
  p1,
  SavePlotFile,
  EdgeCut = 0.05,
  VertexLabel = TRUE,
  AddTrait = TRUE,
  WeightedEdge = TRUE,
  ShowNodes = FALSE,
  PlotTitle = ""
)
}
\description{
(INTERNAL)
}
\keyword{internal}
