% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAdiagTools.R
\name{PlotNetwork2}
\alias{PlotNetwork2}
\title{(INTERNAL)}
\usage{
PlotNetwork2(
  EdgeMatrix,
  CorrMatrix,
  mRNAmiRNAmodules,
  ModuleIdx,
  SaveFile,
  CorM = NULL,
  EdgeCut = 0.05,
  VertexLabel = TRUE,
  AddTrait = FALSE,
  WeightedEdge = FALSE,
  ShowNodes = TRUE,
  PlotTitle = ""
)
}
\description{
(INTERNAL)
}
\keyword{internal}
