% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAdiagTools.R
\name{PlotReducedNetwork}
\alias{PlotReducedNetwork}
\title{(INTERNAL)}
\usage{
PlotReducedNetwork(
  EdgeMatrix,
  CorrMatrix,
  mRNAmiRNAmodules,
  ModuleIdx,
  AddCorrSign = TRUE,
  p1 = 3212,
  SaveFile,
  CorM = NULL,
  EdgeCut = 0.05,
  VertexLabel = TRUE,
  AddTrait = FALSE,
  WeightedEdge = FALSE,
  ShowNodes = TRUE,
  PlotTitle = ""
)
}
\description{
(INTERNAL)
}
\keyword{internal}
