% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCAdiagTools.R
\name{RunWGCNA}
\alias{RunWGCNA}
\title{(INTERNAL)}
\usage{
RunWGCNA(
  X,
  wgcnaResult,
  wgcnaPlot,
  Power = 6,
  Network = "unsigned",
  Corr = "cor",
  minModuleSize = 3,
  mergeCutHeight = 0.25
)
}
\description{
(INTERNAL)
}
\keyword{internal}
