% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetTools.R
\name{SmCCNet_Running}
\alias{SmCCNet_Running}
\title{Running SmCCNet with criterion specified by users}
\usage{
SmCCNet_Running(
  dir,
  X1,
  X2,
  Y,
  factor_1,
  factor_2,
  factor_3,
  pheno,
  K,
  s1 = 0.7,
  s2 = 0.9,
  subSamp = 1000,
  pen1 = seq(0.05, 0.3, by = 0.025),
  pen2 = seq(0.05, 0.3, by = 0.025),
  Type = "PSOCK"
)
}
\arguments{
\item{dir}{the working directory}

\item{X1}{first dataset}

\item{X2}{second dataset}

\item{Y}{phenotype}

\item{factor_1}{scaling factor 1}

\item{factor_2}{scaling factor 2}

\item{factor_3}{scaling factor 3}

\item{pheno}{name of phenotype}

\item{K}{K-folds cross-validation}

\item{s1}{subsampling parameter 1}

\item{s2}{subsampling parameter 2}

\item{subSamp}{subsampling times}

\item{pen1}{first penalty sequence}

\item{pen2}{second penalty sequence}

\item{Type}{parallel processing type}
}
\value{
SmCCNet result stored in the local directory
}
\description{
Running SmCCNet with criterion specified by users
}
